requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
Django==4.2.18
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
